Update test case
