+++ tests/api/health_check.rs
@@ -1,4 +1,4 @@
-use std::net::TcpListener;
+use crate::helpers::spawn_app;
 
 #[tokio::test]
 async fn health_check_works() {
@@ -16,13 +16,4 @@ async fn health_check_works() {
     // Assert
     assert!(response.status().is_success());
     assert_eq!(Some(0), response.content_length());
-}
-
-fn spawn_app() -> String {
-    let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind random port");
-    let port = listener.local_addr().unwrap().port();
-    let server = pharmacity::startup::run(listener).expect("Failed to bind address");
-    let _ = tokio::spawn(server);
-
-    format!("http://127.0.0.1:{port}")
 }
