+++ tests/api/users.rs
@@ -1,0 +1,7 @@
+use crate::helpers::spawn_app;
+
+#[tokio::test]
+async fn post_user_returns_a_200_for_valid() {
+    // Arrange
+    let app_address = spawn_app();
+}
